+++ src/types/index.ts
@@ -9,3 +9,4 @@ export * from "./datafeed";
 export * from "./linkage";
 export * from "./macro";
 export * from "./adapter";
+export * from "./filters";